+++ src/Service.php
@@ -21,6 +21,7 @@ use Embed\Http\Crawler;
 use Embed\Http\CurlClient;
 use Embed\Http\Url;
 use spicyweb\embeddedassets\adapters\akamai\Extractor as AkamaiExtractor;
+use spicyweb\embeddedassets\adapters\bluesky\Extractor as BlueskyExtractor;
 use spicyweb\embeddedassets\adapters\default\Extractor as DefaultExtractor;
 use spicyweb\embeddedassets\adapters\default\detectors\Type as TypeDetector;
 use spicyweb\embeddedassets\adapters\googlemaps\Extractor as GoogleMapsExtractor;
@@ -116,6 +117,8 @@ class Service extends Component
         $clientSettings = [];
         $adapters = [
             'akamaized.net' => AkamaiExtractor::class,
+            'bsky.app' => BlueskyExtractor::class,
+            'bsky.social' => BlueskyExtractor::class,
             'pbs.org' => PbsExtractor::class,
             'nhpbs.org' => PbsExtractor::class,
             'openstreetmap.org' => OpenStreetMapExtractor::class,