+++ CHANGELOG.md
@@ -3,12 +3,14 @@
 ## Unreleased
 
 ### Added
+- Added `spicyweb\embeddedassets\adapters\bluesky\Extractor`
+- Added `spicyweb\embeddedassets\adapters\bluesky\OEmbed`
 - Added `spicyweb\embeddedassets\adapters\openstreetmap\Extractor`
 - Added `spicyweb\embeddedassets\adapters\openstreetmap\detectors\Code`
 - Added `spicyweb\embeddedassets\adapters\openstreetmap\detectors\Type`
 
 ### Changed
-- The default value for `spicyweb\embeddedassets\models\Settings::$whitelist` now includes `'openstreetmap.org'`
+- The default value for `spicyweb\embeddedassets\models\Settings::$whitelist` now includes `'bsky.app'`, `'bsky.social'`, and `'openstreetmap.org'`
 
 ## 4.0.1 - 2024-03-28